+++ checkrules.js
@@ -43,7 +43,7 @@ const
       'package': 'eslint-plugin-react',
       'docs': 'https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/{rule}.md',
       'file': 'react.js',
-      'prefix': 'react/',
+      'namespace': 'react/',
       'processRemoteData': processReactPlugin
     }
   ],
@@ -57,7 +57,7 @@ const
         config = require(`./${ruleSource.file}`)
 
       localRules = Object.keys(config.rules)
-        .filter(r => !ruleSource.prefix || null !== r.match(ruleSource.prefix))
+        .filter(r => !ruleSource.namespace || null !== r.match(ruleSource.namespace))
     }
     catch(_) {
       // noop
@@ -97,7 +97,7 @@ const
   findNewRules = ruleSource => {
     const
       newRules = ruleSource.remoteRules
-        .filter(r => !ruleSource.localRules.includes(`${ruleSource.prefix || ''}${r}`))
+        .filter(r => !ruleSource.localRules.includes(`${ruleSource.namespace || ''}${r}`))
         .map(r => `   ${chalk.green('+')} \
 ${chalk.cyan(ruleSource.docs.replace(/{rule}/, chalk.bold(r)))}`)
 
@@ -107,7 +107,7 @@ ${chalk.cyan(ruleSource.docs.replace(/{rule}/, chalk.bold(r)))}`)
   findRemovedRules = ruleSource => {
     const
       removedRules = ruleSource.localRules
-        .filter(r => !ruleSource.remoteRules.includes(r.replace(ruleSource.prefix || '', '')))
+        .filter(r => !ruleSource.remoteRules.includes(r.replace(ruleSource.namespace || '', '')))
         .map(r => `   ${chalk.red.dim('-')} ${chalk.gray(r)}`)
 
     return Object.assign({}, ruleSource, { removedRules })